refactor: simplify reindex code after review
- Extract `yield_to_search()` helper to deduplicate 3 spin-wait loops
- Replace Ping+Reindex double IPC in cmd_dict_update with single
  try_send_request(Reindex) call
- Change ReindexStatus.errors from usize to i64 for type consistency

+++ src/status.rs
@@ -53,7 +53,7 @@ pub struct ReindexStatus {
     pub kind: ReindexKind,
     pub total: i64,
     pub processed: i64,
-    pub errors: usize,
+    pub errors: i64,
     pub started_at: String,
 }
